refactored the oputs tb methods into the fucntions.outs module. added new public method for access to them. updated texas holdem class to refer to new public method unit tests top follow
