+++ src/pypoker/engine/hand_solver/texas_holdem.py
@@ -8,13 +8,12 @@ from pypoker.engine.hand_solver.constants import (
     HAND_RANK,
     BEST_HAND,
     HAND_DESCRIPTION,
-    OUTS_TB_METHOD,
     TIEBREAKER,
     OUT_STRING, HAND_TYPE_STRAIGHT_FLUSH, HAND_TYPE_QUADS, HAND_TYPE_FULL_HOUSE, HAND_TYPE_FLUSH, HAND_TYPE_STRAIGHT,
     HAND_TYPE_TRIPS, HAND_TYPE_TWO_PAIR, HAND_TYPE_PAIR, HAND_TYPE_HIGH_CARD, GAME_TYPE_TEXAS_HOLDEM,
 )
 from pypoker.engine.hand_solver.functions import hand_test, rank_hand_type, describe_hand, find_outs_scenarios
-from pypoker.engine.hand_solver.functions.outs import claim_out_string
+from pypoker.engine.hand_solver.functions.outs import claim_out_string, tiebreak_outs_draw
 from pypoker.engine.hand_solver.utils import get_all_combinations
 
 
@@ -42,51 +41,15 @@ class TexasHoldemHandSolver(BaseHandSolver):
 
     def __init__(self):
         self._hand_rankings = [
-            {
-                HAND_TITLE: HAND_TYPE_STRAIGHT_FLUSH,
-                HAND_RANK: 1,
-                OUTS_TB_METHOD: self._outs_tb_straight_flush,
-            },
-            {
-                HAND_TITLE: HAND_TYPE_QUADS,
-                HAND_RANK: 2,
-                OUTS_TB_METHOD: self._outs_tb_quads,
-            },
-            {
-                HAND_TITLE: HAND_TYPE_FULL_HOUSE,
-                HAND_RANK: 3,
-                OUTS_TB_METHOD: self._outs_tb_full_house,
-            },
-            {
-                HAND_TITLE: HAND_TYPE_FLUSH,
-                HAND_RANK: 4,
-                OUTS_TB_METHOD: self._outs_tb_flush,
-            },
-            {
-                HAND_TITLE: HAND_TYPE_STRAIGHT,
-                HAND_RANK: 5,
-                OUTS_TB_METHOD: self._outs_tb_straight,
-            },
-            {
-                HAND_TITLE: HAND_TYPE_TRIPS,
-                HAND_RANK: 6,
-                OUTS_TB_METHOD: self._outs_tb_trips,
-            },
-            {
-                HAND_TITLE: HAND_TYPE_TWO_PAIR,
-                HAND_RANK: 7,
-                OUTS_TB_METHOD: self._outs_tb_two_pair,
-            },
-            {
-                HAND_TITLE: HAND_TYPE_PAIR,
-                HAND_RANK: 8,
-                OUTS_TB_METHOD: self._outs_tb_pair,
-            },
-            {
-                HAND_TITLE: HAND_TYPE_HIGH_CARD,
-                HAND_RANK: 9,
-                OUTS_TB_METHOD: self._outs_tb_high_card,
-            },
+            {HAND_TITLE: HAND_TYPE_STRAIGHT_FLUSH, HAND_RANK: 1},
+            {HAND_TITLE: HAND_TYPE_QUADS, HAND_RANK: 2},
+            {HAND_TITLE: HAND_TYPE_FULL_HOUSE, HAND_RANK: 3},
+            {HAND_TITLE: HAND_TYPE_FLUSH, HAND_RANK: 4},
+            {HAND_TITLE: HAND_TYPE_STRAIGHT, HAND_RANK: 5},
+            {HAND_TITLE: HAND_TYPE_TRIPS, HAND_RANK: 6},
+            {HAND_TITLE: HAND_TYPE_TWO_PAIR, HAND_RANK: 7},
+            {HAND_TITLE: HAND_TYPE_PAIR, HAND_RANK: 8},
+            {HAND_TITLE: HAND_TYPE_HIGH_CARD, HAND_RANK: 9}
         ]
 
     ########################
@@ -288,8 +251,8 @@ class TexasHoldemHandSolver(BaseHandSolver):
                             hole_cards[player] = player_hole_cards[player]
 
                         drawn_cards = [Card(card_id) for card_id in out_ids]
-                        winner = self._hand_rankings[rank - 1][OUTS_TB_METHOD](
-                            tiebreakers, hole_cards, board_cards, drawn_cards
+                        winner = tiebreak_outs_draw(GAME_TYPE_TEXAS_HOLDEM, self._hand_rankings[HAND_TITLE],
+                            tiebreakers=tiebreakers, hole_cards=hole_cards, board_cards=board_cards, drawn_cards=drawn_cards
                         )
                         if winner not in wins.keys():
                             wins[winner] = 1
@@ -309,434 +272,6 @@ class TexasHoldemHandSolver(BaseHandSolver):
         }
         return odds
 
-    #############################
-    #  OUTS TIEBREAKER METHODS  #
-    ######################1#######
-
-    @staticmethod
-    def _outs_tb_straight_flush(
-        tiebreakers: Dict,
-        hole_cards: Dict[str, List[Card]],
-        board_cards: List[Card],
-        drawn_cards: List[Card],
-    ) -> str:
-        """
-        Private method to determine which player would have the stronger straight flush hand. given the drawn cards.
-
-        :param tiebreakers: dictionary of each players tiebreak information for this draw.
-        :param hole_cards: dictionary of each players hole cards
-        :param board_cards: List of the board cards that have been dealt
-        :param drawn_cards: List of cards that would be drawn in this scenario
-
-        :return: Name of the winning player. if a tie occurs, then return a TIE(<PLAYER_NAMES>) where PLAYER_NAMES is
-                 a comma-separated list of the players who have tied
-        """
-
-        max_tiebreaker = max([tiebreaker for tiebreaker in tiebreakers.values()])
-        winners = [
-            player
-            for player, tiebreaker in tiebreakers.items()
-            if tiebreaker == max_tiebreaker
-        ]
-        return winners[0] if len(winners) == 1 else f"TIE({','.join(winners)})"
-
-    @staticmethod
-    def _outs_tb_quads(
-        tiebreakers: Dict,
-        hole_cards: Dict[str, List[Card]],
-        board_cards: List[Card],
-        drawn_cards: List[Card],
-    ) -> str:
-        """
-        Private method to determine which player would have the stronger quads hand. given the drawn cards.
-
-        :param tiebreakers: dictionary of each players tiebreak information for this draw.
-        :param hole_cards: dictionary of each players hole cards
-        :param board_cards: List of the board cards that have been dealt
-        :param drawn_cards: List of cards that would be drawn in this scenario
-
-        :return: Name of the winning player. if a tie occurs, then return a TIE(<PLAYER_NAMES>) where PLAYER_NAMES is
-                 a comma-separated list of the players who have tied
-        """
-
-        max_tiebreaker = max([tiebreaker for tiebreaker in tiebreakers.values()])
-        possible_winners = [
-            player
-            for player, tiebreaker in tiebreakers.items()
-            if tiebreaker == max_tiebreaker
-        ]
-
-        if len(possible_winners) == 1:
-            return possible_winners[0]
-
-        player_kickers = {}
-        for player in possible_winners:
-            player_values = [card.value for card in hole_cards[player]]
-            player_values.extend([card.value for card in board_cards])
-            player_values.extend([card.value for card in drawn_cards])
-            player_values = [
-                value for value in player_values if value != max_tiebreaker
-            ]
-
-            player_kickers[player] = max(player_values)
-
-        max_kicker = max([kicker for kicker in player_kickers.values()])
-        winners = [
-            player for player, kicker in player_kickers.items() if kicker == max_kicker
-        ]
-        return winners[0] if len(winners) == 1 else f"TIE({','.join(winners)})"
-
-    @staticmethod
-    def _outs_tb_full_house(
-        tiebreakers: Dict,
-        hole_cards: Dict[str, List[Card]],
-        board_cards: List[Card],
-        drawn_cards: List[Card],
-    ) -> str:
-        """
-        Private method to determine which player would have the stronger full house hand. given the drawn cards.
-
-        :param tiebreakers: dictionary of each players tiebreak information for this draw.
-        :param hole_cards: dictionary of each players hole cards
-        :param board_cards: List of the board cards that have been dealt
-        :param drawn_cards: List of cards that would be drawn in this scenario
-
-        :return: Name of the winning player. if a tie occurs, then return a TIE(<PLAYER_NAMES>) where PLAYER_NAMES is
-                 a comma-separated list of the players who have tied
-        """
-
-        max_trips = max([tiebreaker[0] for tiebreaker in tiebreakers.values()])
-        max_trips_players = [
-            player
-            for player, tiebreaker in tiebreakers.items()
-            if tiebreaker[0] == max_trips
-        ]
-
-        if len(max_trips_players) == 1:
-            return max_trips_players[0]
-
-        max_pair = max(
-            [
-                tiebreaker[1]
-                for player, tiebreaker in tiebreakers.items()
-                if player in max_trips_players
-            ]
-        )
-        max_pair_players = [
-            player
-            for player, tiebreaker in tiebreakers.items()
-            if tiebreaker[1] == max_pair and player in max_trips_players
-        ]
-
-        return (
-            max_pair_players[0]
-            if len(max_pair_players) == 1
-            else f"TIE({','.join(max_pair_players)})"
-        )
-
-    @staticmethod
-    def _outs_tb_flush(
-        tiebreakers: Dict,
-        hole_cards: Dict[str, List[Card]],
-        board_cards: List[Card],
-        drawn_cards: List[Card],
-    ) -> str:
-        """
-        Private method to determine which player would have the stronger flush hand. given the drawn cards.
-
-        :param tiebreakers: dictionary of each players tiebreak information for this draw.
-        :param hole_cards: dictionary of each players hole cards
-        :param board_cards: List of the board cards that have been dealt
-        :param drawn_cards: List of cards that would be drawn in this scenario
-
-        :return: Name of the winning player. if a tie occurs, then return a TIE(<PLAYER_NAMES>) where PLAYER_NAMES is
-                 a comma-separated list of the players who have tied
-        """
-
-        player_flush_values = {}
-        for player, cards in hole_cards.items():
-            flush_values = [
-                card.value for card in cards if card.suit == tiebreakers[player]
-            ]
-            flush_values.extend(
-                [card.value for card in board_cards if card.suit == tiebreakers[player]]
-            )
-            flush_values.extend(
-                [card.value for card in drawn_cards if card.suit == tiebreakers[player]]
-            )
-
-            player_flush_values[player] = sorted(flush_values, reverse=True)
-
-        winners = [player for player in hole_cards.keys()]
-
-        for index in range(5):
-            max_index_value = max(
-                [
-                    values[index]
-                    for player, values in player_flush_values.items()
-                    if player in winners
-                ]
-            )
-
-            winners = [
-                player
-                for player in winners
-                if player_flush_values[player][index] == max_index_value
-            ]
-
-        return winners[0] if len(winners) == 1 else f"TIE({','.join(winners)})"
-
-    @staticmethod
-    def _outs_tb_straight(
-        tiebreakers: Dict,
-        hole_cards: Dict[str, List[Card]],
-        board_cards: List[Card],
-        drawn_cards: List[Card],
-    ) -> str:
-        """
-        Private method to determine which player would have the stronger straight hand. given the drawn cards.
-
-        :param tiebreakers: dictionary of each players tiebreak information for this draw.
-        :param hole_cards: dictionary of each players hole cards
-        :param board_cards: List of the board cards that have been dealt
-        :param drawn_cards: List of cards that would be drawn in this scenario
-
-        :return: Name of the winning player. if a tie occurs, then return a TIE(<PLAYER_NAMES>) where PLAYER_NAMES is
-                 a comma-separated list of the players who have tied
-        """
-
-        max_tiebreaker = max([tiebreaker for tiebreaker in tiebreakers.values()])
-        winners = [
-            player
-            for player, tiebreaker in tiebreakers.items()
-            if tiebreaker == max_tiebreaker
-        ]
-        return winners[0] if len(winners) == 1 else f"TIE({','.join(winners)})"
-
-    @staticmethod
-    def _outs_tb_trips(
-        tiebreakers: Dict,
-        hole_cards: Dict[str, List[Card]],
-        board_cards: List[Card],
-        drawn_cards: List[Card],
-    ) -> str:
-        """
-        Private method to determine which player would have the stronger trips hand. given the drawn cards.
-
-        :param tiebreakers: dictionary of each players tiebreak information for this draw.
-        :param hole_cards: dictionary of each players hole cards
-        :param board_cards: List of the board cards that have been dealt
-        :param drawn_cards: List of cards that would be drawn in this scenario
-
-        :return: Name of the winning player. if a tie occurs, then return a TIE(<PLAYER_NAMES>) where PLAYER_NAMES is
-                 a comma-separated list of the players who have tied
-        """
-
-        max_tiebreaker = max([tiebreaker for tiebreaker in tiebreakers.values()])
-        possible_winners = [
-            player
-            for player, tiebreaker in tiebreakers.items()
-            if tiebreaker == max_tiebreaker
-        ]
-
-        if len(possible_winners) == 1:
-            return possible_winners[0]
-
-        player_kickers = {}
-        for player in possible_winners:
-            player_values = [card.value for card in hole_cards[player]]
-            player_values.extend([card.value for card in board_cards])
-            player_values.extend([card.value for card in drawn_cards])
-            player_values = [
-                value for value in player_values if value != max_tiebreaker
-            ]
-
-            player_kickers[player] = sorted(player_values, reverse=True)
-
-        winners = [player for player in possible_winners]
-        for index in range(2):
-            max_index_value = max(
-                [
-                    values[index]
-                    for player, values in player_kickers.items()
-                    if player in winners
-                ]
-            )
-
-            winners = [
-                player
-                for player in winners
-                if player_kickers[player][index] == max_index_value
-            ]
-
-        return winners[0] if len(winners) == 1 else f"TIE({','.join(winners)})"
-
-    @staticmethod
-    def _outs_tb_two_pair(
-        tiebreakers: Dict,
-        hole_cards: Dict[str, List[Card]],
-        board_cards: List[Card],
-        drawn_cards: List[Card],
-    ) -> str:
-        """
-        Private method to determine which player would have the stronger two pair hand. given the drawn cards.
-
-        :param tiebreakers: dictionary of each players tiebreak information for this draw.
-        :param hole_cards: dictionary of each players hole cards
-        :param board_cards: List of the board cards that have been dealt
-        :param drawn_cards: List of cards that would be drawn in this scenario
-
-        :return: Name of the winning player. if a tie occurs, then return a TIE(<PLAYER_NAMES>) where PLAYER_NAMES is
-                 a comma-separated list of the players who have tied
-        """
-
-        max_high_pair = max([tiebreaker[0] for tiebreaker in tiebreakers.values()])
-        max_high_pair_players = [
-            player
-            for player, tiebreaker in tiebreakers.items()
-            if tiebreaker[0] == max_high_pair
-        ]
-
-        if len(max_high_pair_players) == 1:
-            return max_high_pair_players[0]
-
-        max_low_pair = max(
-            [
-                tiebreaker[1]
-                for player, tiebreaker in tiebreakers.items()
-                if player in max_high_pair_players
-            ]
-        )
-        max_low_pair_players = [
-            player
-            for player, tiebreaker in tiebreakers.items()
-            if tiebreaker[1] == max_low_pair and player in max_high_pair_players
-        ]
-
-        if len(max_low_pair_players) == 1:
-            return max_low_pair_players[0]
-
-        player_kickers = {}
-        for player in max_low_pair_players:
-            player_values = [card.value for card in hole_cards[player]]
-            player_values.extend([card.value for card in board_cards])
-            player_values.extend([card.value for card in drawn_cards])
-            player_values = [
-                value
-                for value in player_values
-                if value not in [max_high_pair, max_low_pair]
-            ]
-
-            player_kickers[player] = max(player_values)
-
-        max_kicker = max([kicker for kicker in player_kickers.values()])
-        winners = [
-            player for player, kicker in player_kickers.items() if kicker == max_kicker
-        ]
-        return winners[0] if len(winners) == 1 else f"TIE({','.join(winners)})"
-
-    @staticmethod
-    def _outs_tb_pair(
-        tiebreakers: Dict,
-        hole_cards: Dict[str, List[Card]],
-        board_cards: List[Card],
-        drawn_cards: List[Card],
-    ) -> str:
-        """
-        Private method to determine which player would have the stronger pair hand. given the drawn cards.
-
-        :param tiebreakers: dictionary of each players tiebreak information for this draw.
-        :param hole_cards: dictionary of each players hole cards
-        :param board_cards: List of the board cards that have been dealt
-        :param drawn_cards: List of cards that would be drawn in this scenario
-
-        :return: Name of the winning player. if a tie occurs, then return a TIE(<PLAYER_NAMES>) where PLAYER_NAMES is
-                 a comma-separated list of the players who have tied
-        """
-
-        max_tiebreaker = max([tiebreaker for tiebreaker in tiebreakers.values()])
-        possible_winners = [
-            player
-            for player, tiebreaker in tiebreakers.items()
-            if tiebreaker == max_tiebreaker
-        ]
-
-        if len(possible_winners) == 1:
-            return possible_winners[0]
-
-        player_kickers = {}
-        for player in possible_winners:
-            player_values = [card.value for card in hole_cards[player]]
-            player_values.extend([card.value for card in board_cards])
-            player_values.extend([card.value for card in drawn_cards])
-            player_values = [
-                value for value in player_values if value != max_tiebreaker
-            ]
-
-            player_kickers[player] = sorted(player_values, reverse=True)
-
-        winners = [player for player in possible_winners]
-        for index in range(3):
-            max_index_value = max(
-                [
-                    values[index]
-                    for player, values in player_kickers.items()
-                    if player in winners
-                ]
-            )
-
-            winners = [
-                player
-                for player in winners
-                if player_kickers[player][index] == max_index_value
-            ]
-
-        return winners[0] if len(winners) == 1 else f"TIE({','.join(winners)})"
-
-    @staticmethod
-    def _outs_tb_high_card(
-        tiebreakers: Dict,
-        hole_cards: Dict[str, List[Card]],
-        board_cards: List[Card],
-        drawn_cards: List[Card],
-    ) -> str:
-        """
-        Private method to determine which player would have the stronger high card hand. given the drawn cards.
-
-        :param tiebreakers: dictionary of each players tiebreak information for this draw.
-        :param hole_cards: dictionary of each players hole cards
-        :param board_cards: List of the board cards that have been dealt
-        :param drawn_cards: List of cards that would be drawn in this scenario
-
-        :return: Name of the winning player. if a tie occurs, then return a TIE(<PLAYER_NAMES>) where PLAYER_NAMES is
-                 a comma-separated list of the players who have tied
-        """
-
-        player_kickers = {}
-        for player in hole_cards.keys():
-            player_values = [card.value for card in hole_cards[player]]
-            player_values.extend([card.value for card in board_cards])
-            player_values.extend([card.value for card in drawn_cards])
-
-            player_kickers[player] = sorted(player_values, reverse=True)
-
-        winners = [player for player in tiebreakers.keys()]
-        for index in range(5):
-            max_index_value = max(
-                [
-                    values[index]
-                    for player, values in player_kickers.items()
-                    if player in winners
-                ]
-            )
-
-            winners = [
-                player
-                for player in winners
-                if player_kickers[player][index] == max_index_value
-            ]
-
-        return winners[0] if len(winners) == 1 else f"TIE({','.join(winners)})"
 
     #########################
     #  MISC HELPER METHODS  #
